Remove extra print statements

+++ src/service/helper.py
@@ -3,10 +3,11 @@ import shutil
 
 
 def clean_up(path):
-    if os.path.isfile(path):
-        print(f' Removing File: {path}')
-        os.remove(path)
-    else:
-        folder = os.path.join(path)
-        print(f' Removing Folder: {path}')
-        shutil.rmtree(folder, ignore_errors=False)
+    if os.path.exists(path):
+        if os.path.isfile(path):
+            print(f' Removing File: {path}')
+            os.remove(path)
+        else:
+            folder = os.path.join(path)
+            print(f' Removing Folder: {path}')
+            shutil.rmtree(folder, ignore_errors=False)
